Add files via upload
Our product is Sashakt; a software with an independent behaviour. This software deals with the tedious task of case lodging as well as case monitoring. the self monitoring phase helps a time bound evaluation each and every phase of a case. The product is new and underdevelopment due to which it could not be completed. But our product has a future which would evolutionallise our way of dealing with crimes.